requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
deep-translator==1.11.4
defusedxml==0.7.1
distlib==0.3.7
distro==1.8.0
dj-rest-auth==4.0.1
Django==3.2.20
django-allauth==0.54.0
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
filelock==3.12.2
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
joblib==1.2.0
langdetect==1.0.9
lightgbm==4.0.0
lz4==4.3.2
numpy==1.25.2
oauthlib==3.2.2
openai==1.3.5
Pillow==10.0.0
pipenv==2023.7.23
platformdirs==3.9.1
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python3-openid==3.2.0
pytz==2023.3
q==2.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
threadpoolctl==3.1.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.2
virtualenv-clone==0.5.7
xgboost==1.7.6
